requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
requests
python-multipart
